+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## Neste versjon
 
+- ✨ **Feedback**. Brukere kan gi tilbakemeldinger om ideer og bugs.
+
 ## Versjon 2024.18.09
 
 - 🦟 **Safari zoom**. En bug der safari på mobil zoomer inn på input er nå fikset.